Add single file string builder implementation
This is my attempt at a single header string builder library. To use it,
define the macro

    STRING_BUILDER_IMPLEMENTATION

_exactly once_ before including the header. The design of the library is
heavily inspired by the stb libraries - if you're familiar with them,
you should have no problems using this library.

The file currently depends on a few functions from libc which I aim to
replace in the future for a fully standalone library.

I'm almost certain there are quite a few bugs in this. Buyer beware!