Bundle up some utf8 utility functions into a header
Provide the functions encode_codepoint(), which takes a Unicode
codepoint and encodes it into valid UTF-8, peek_codepoint(), which
returns the next Unicode codepoint in a UTF-8 encoded string and
advance_codepoint(), which not only returns the next codepoint but also
eats it and advances the provided string.

Structure the header like we did with the string builder. A bit overkill
for three functions, but this thing might grow a little.