Add unchecked_append()
I've seen this in SerenityOS and it makes sense - if we know for sure
that the current bucket is big enough to hold the string we want to
append, we don't need to check.

This leads to a bit of code duplication down the road, if we want to add
a unchecked_appendf() version. Might have to refactor the functions a
bit. We'll see.