Add small unit test header
I'm usually not a big proponent of unit tests, but for stuff like these
single header libraries, it makes sense. And we already have some hacked
together tests for the string builder library, which are barely readable
and not reusable at all for other libraries. As this repository grows
and other headers are added, this would become a giant mess.

This header aims at making things a bit tidier and giving tests a
uniform structure. Right now, there are a lot of things missing. Like
with all the other libraries, start small, grow where needed.