Let the append functions take a void * instead of char *
If we're going to allow arbitrary data in the string builder, we ought
to be consistent with it. The append function has been split up into a
general function, which takes every type of pointer, and a string
function, which specfically takes a string.

This also has the nice effect of cleaning up the code a little.