Split up to_string() into to_c_string() and to_byte_string()
Since we're having different functions for appending C strings and raw
bytes, let's also have different functions for outputting those.

- to_c_string() takes a char ** and returns the contents of the string
  builder with a terminating NULL character.

- to_byte_string() takes an unsigned char ** and outputs string
  builder's contents without a NULL character at the end.

It would probably be nice if we were returning the string in the string
builder at the first found NULL in to_c_string(), but I don't think it's
that important.