Add cpuid abstraction header
This header provides a uniform way across different compilers (namely
MSVC and GCC/Clang) to call the __cpuid intrinsic and defines multiple
macros to extract information out of the returning values.

This can be used to check if the CPU the program is running on supports
a specific instruction extension. It also provides a lot of other
information about the CPU.

I did not check every provided macro for validity - in fact, I only
tested a few, so no guarantees are given. I'm sure there are quite a few
bugs in here. I'm not going to be able to find them all, given that I
only have a Kaby Lake CPU to test the header on at the moment.